Cleanup & documentation

+++ RobotControl/src/de/naoth/rc/tools/CommandRecorder.java
@@ -46,47 +46,46 @@ public class CommandRecorder implements Plugin, ConnectionStatusListener, Object
     @InjectPlugin
     public static SwingCommandExecutor commandExecutor;
     
-    
     private final String recordingPrefix = "recording_";
     private final String recordingSuffix = ".bin";
+    private final String menuMacrosTooltip = "Not connected to robot";
+    private final String menuRecordingStart = "Start Recording";
+    private final String menuRecordingStop = "Stop Recording";
     
     private boolean isRecording = false;
+    private final List<Command> log = new LinkedList<>();
     
-    private List<Command> log = new LinkedList<>();
-    
-    private JMenu menuMacros = new javax.swing.JMenu("Macros");
-    private JMenuItem menuStartRecording = new JMenuItem("Start Recording");
+    private final JMenu menuMacros = new javax.swing.JMenu("Macros");
+    private final JMenuItem menuStartRecording = new JMenuItem(menuRecordingStart);
 
     /**
      * Gets called, when the RC plugin was loaded.
-     * Adds the icon/label to the RC statusbar and registers itself as connection listener.
+     * Adds the "Macros" menu item to the RC menubar and registers itself as connection listener.
      * 
-     * @param robotControl 
+     * @param robotControl  the RC instance (JFrame)
      */
     @PluginLoaded
     public void loaded(RobotControl robotControl) {
         rc.getMessageServer().addConnectionStatusListener(this);
-
-        menuStartRecording.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                handleRecording();
-            }
-        });
+        
+        menuMacros.setEnabled(false);
+        menuMacros.setToolTipText(menuMacrosTooltip);
+        menuMacros.setHorizontalTextPosition(SwingConstants.LEADING);
         menuMacros.add(menuStartRecording);
         menuMacros.addSeparator();
-        menuMacros.setHorizontalTextPosition(SwingConstants.LEADING);
-        menuMacros.setEnabled(false);
-        menuMacros.setToolTipText("Not connected to robot");
 
-        Arrays.asList((new File(robotControl.getConfigPath())).listFiles((dir, name) -> {
-            return name.startsWith(recordingPrefix) && name.endsWith(recordingSuffix);
-        })).forEach((f) -> {
-            String name = f.getName().substring(recordingPrefix.length(), f.getName().length() - recordingSuffix.length());
-            addMenuItem(name);
-        });
-
-        JMenuBar menu = ((RobotControlImpl)robotControl).getJMenuBar();
+        menuStartRecording.addActionListener((ActionEvent e) -> { handleRecording(); });
         
+        addRecordingsToMenu();
+        insertMacrosMenuToMenuBar();
+    }
+    
+    /**
+     * Inserts the "Macros" menu item to the end of the left menubar part.
+     */
+    private void insertMacrosMenuToMenuBar() {
+        JMenuBar menu = ((RobotControlImpl)robotControl).getJMenuBar();
+        // find the index of the 'Filler', in order to insert the menu before it
         int idx = menu.getComponentCount();
         for (Component component : menu.getComponents()) {
             if (component instanceof Box.Filler) {
@@ -95,7 +94,25 @@ public class CommandRecorder implements Plugin, ConnectionStatusListener, Object
         }
         menu.add(menuMacros, idx);
     }
+    
+    /**
+     * Searches for recoding files and adds them to the macro menu.
+     */
+    private void addRecordingsToMenu() {
+        Arrays.asList((new File(robotControl.getConfigPath())).listFiles((dir, name) -> {
+            return name.startsWith(recordingPrefix) && name.endsWith(recordingSuffix);
+        })).forEach((f) -> {
+            String name = f.getName().substring(recordingPrefix.length(), f.getName().length() - recordingSuffix.length());
+            addMenuItem(name);
+        });
+    }
 
+    /**
+     * Adds a recording to the macros menu.
+     * Also sets the tooltip and handles the click events.
+     * 
+     * @param name  the recording name
+     */
     private void addMenuItem(String name) {
         JMenuItem item = new JMenuItem(name);
         item.setToolTipText("<html>Replays the recorded commands.<br>Use <i>Ctrl+Click</i> to delete this recording.</html>");
@@ -103,7 +120,11 @@ public class CommandRecorder implements Plugin, ConnectionStatusListener, Object
             JMenuItem source = (JMenuItem) e.getSource();
             if((e.getModifiers() & ActionEvent.CTRL_MASK) == ActionEvent.CTRL_MASK) {
                 // delete requested
-                if(JOptionPane.showConfirmDialog(((RobotControlImpl)robotControl), "Do you want to remove the recording '"+name+"'?", "Remove recording", JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION) {
+                if(JOptionPane.showConfirmDialog(
+                        ((RobotControlImpl)robotControl), 
+                        "Do you want to remove the recording '"+name+"'?", 
+                        "Remove recording", 
+                        JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION) {
                     if(recordingFileName(name).delete()) { menuMacros.remove(source); }
                 }
             } else {
@@ -116,13 +137,20 @@ public class CommandRecorder implements Plugin, ConnectionStatusListener, Object
                         Logger.getLogger(RobotControlImpl.class.getName()).log(Level.SEVERE, null, ex);
                     }
                 } else {
-                    JOptionPane.showMessageDialog(((RobotControlImpl)robotControl), "The '"+source.getText()+"' recording file doesn't exists!?", "Missing recording file", JOptionPane.ERROR_MESSAGE);
+                    JOptionPane.showMessageDialog(
+                            ((RobotControlImpl)robotControl), 
+                            "The '"+source.getText()+"' recording file doesn't exists!?", 
+                            "Missing recording file", 
+                            JOptionPane.ERROR_MESSAGE);
                 }
             }
         });
         menuMacros.add(item);
     }
     
+    /**
+     * Starts or stops the recording, based on the current state (helper method).
+     */
     private void handleRecording() {
         if (isRecording) {
             stopRecording();
@@ -131,33 +159,52 @@ public class CommandRecorder implements Plugin, ConnectionStatusListener, Object
         }
     }
     
+    /**
+     * Starts the recording.
+     * Registers to the message server and updates the ui.
+     */
     private void startRecording() {
         isRecording = true;
         rc.getMessageServer().addListener(this);
         
         menuMacros.setIcon(new javax.swing.ImageIcon(getClass().getResource("/de/naoth/rc/res/media-record.png")));
-        menuStartRecording.setText("Stop Recording");
+        menuStartRecording.setText(menuRecordingStop);
     }
     
+    /**
+     * Stops the recording.
+     * Unregisters from the message server, updates the ui and saves the recording.
+     */
     private void stopRecording() {
         isRecording = false;
         rc.getMessageServer().removeListener(this);
         
         menuMacros.setIcon(null);
-        menuStartRecording.setText("Start Recording");
+        menuStartRecording.setText(menuRecordingStart);
         
         saveRecording();
     }
     
+    /**
+     * Returns the File for a given recording name (helper method).
+     * 
+     * @param name  the name of the recording
+     * @return      the file object pointing to the file where the recording is/should be saved
+     */
     private File recordingFileName(String name) {
         return new File(robotControl.getConfigPath() + "/" + recordingPrefix + name + recordingSuffix);
     }
     
+    /**
+     * Actually saves the recorded commands.
+     * Therefore checks whether commands were recorded and ask for a name for this
+     * recording. The recorded commands are written as serialized java objects to the file.
+     */
     private void saveRecording() {
         // check if something was recorded
         if (log.isEmpty()) {
             JOptionPane.showMessageDialog(
-                    ((RobotControlImpl)robotControl), 
+                    ((RobotControlImpl)robotControl),
                     "No commands were recorded.", 
                     "Nothing recorded", 
                     JOptionPane.WARNING_MESSAGE);
@@ -174,7 +221,15 @@ public class CommandRecorder implements Plugin, ConnectionStatusListener, Object
         if(name.isEmpty()) { return; }
         // create record file
         File file = recordingFileName(name);
-        // TODO: should the file be overwritten?
+        // if recording name already exists - ask to overwrite
+        if(file.isFile() && JOptionPane.showConfirmDialog(
+                ((RobotControlImpl)robotControl), 
+                "This dialog layout name already exists. Overwrite?",
+                "Overwrite",
+                JOptionPane.YES_NO_OPTION) != JOptionPane.YES_OPTION) {
+            // don't overwrite!
+            return;
+        }
         // write the logged commands to the record file
         try {
             FileOutputStream stream = new FileOutputStream(file);
@@ -193,6 +248,12 @@ public class CommandRecorder implements Plugin, ConnectionStatusListener, Object
         log.clear();
     }
     
+    /**
+     * Loads the recorded commands from the given file.
+     * 
+     * @param f
+     * @throws IOException 
+     */
     private void loadRecording(File f) throws IOException {
         // if not connect, do not replay commands
         if (!rc.getMessageServer().isConnected()) {
@@ -225,26 +286,46 @@ public class CommandRecorder implements Plugin, ConnectionStatusListener, Object
         stream.close();
     }
     
+    /**
+     * Enables the macros menu.
+     * 
+     * @param event 
+     */
     @Override
     public void connected(ConnectionStatusEvent event) {
         menuMacros.setEnabled(true);
         menuMacros.setToolTipText(null);
     }
 
+    /**
+     * Disables the macros menu and sets the tooltip.
+     * @param event 
+     */
     @Override
     public void disconnected(ConnectionStatusEvent event) {
         menuMacros.setEnabled(false);
-        menuMacros.setToolTipText("Not connected to robot");
+        menuMacros.setToolTipText(menuMacrosTooltip);
     }
 
+    /**
+     * Records all 'set' commands.
+     * 
+     * @param cmd 
+     */
     @Override
     public void newObjectReceived(Command cmd) {
         // only store the 'set' commands
         if (cmd.getName().endsWith(":set") || cmd.getName().endsWith(":set_agent")) {
             log.add(cmd);
+            Logger.getLogger(CommandRecorder.class.getName()).log(Level.INFO, "Command logged: " + cmd.getName());
         }
     }
 
+    /**
+     * Ignore any errors.
+     * 
+     * @param cause 
+     */
     @Override
     public void errorOccured(String cause) { /* should never happen! */ }
 }
